+++ filediff-node-module/src/main/java/org/sudu/experiments/js/node/JsSftpClient.java
@@ -70,7 +70,7 @@ public interface JsSftpClient extends JSObject {
   // associated with handle starting at position.
   void write(JSObject handle,
              JsBuffer buffer, int offset, int length,
-             int position,  JsFunctions.Consumer<JSError> cb);
+             double position, JsFunctions.Consumer<JSError> cb);
 
   // Closes the resource associated with handle given by open() or opendir().
   void close(JSObject handle, JsFunctions.Consumer<JSError> cb);
